major update
renamed app to publishingApi
added SPV support
moved node client, wallet client and application API to configurable locations
can use scripts to run more than one node on same machine
publisher registration record publishing is now combined with article record publishing
added support for "sponsored publishing"
fixed record search functions so it properly finds previously published records to avoid unnecessary duplication
significantly simplified the function that handles record publishing
added api endpoint for getWIF
added support for storing various publishing credentials in env file so they don't need to be repeated
